ASDF checkl:tests is now available for non-"formal" tests.
It was pretty annoying to have a file full of named CHECK blocks and
have to run them all manually to redefine them, especially if later
tests run eariler tests.  This is no longer necessary!  Just load
the file with checkl:tests in ASDF, and they'll be defined but not
run.